requirements.txt
astroid==2.3.1
certifi==2019.9.11
colorama==0.4.1
DateTime==4.3
Django==2.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.1
django-filter==2.2.0
django-js-asset==1.2.2
django-tinymce==2.8.0
icalendar==4.0.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
mysqlclient==1.4.4
Pillow==6.2.0
pylint==2.4.2
python-dateutil==2.8.1
pytz==2019.2
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.5
wincertstore==0.2
wrapt==1.11.2
zope.interface==4.6.0
